mcp_client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "MCP Client for Planner AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[client]>=1.6.0",
    "python-dotenv>=1.0.0",
]